Complete second problem in week 1 assignment
"In this problem, your goal is to get used to trees. You will need to read a description of a tree from the input, implement the tree data structure, store the tree and compute its height."

Since this was written in C#, no starter code was provided, so I was required to design an implement a rudimentary tree structure before beginning.

Please note that, due to the submission guidelines, my entire problem solution has to be contained in a single file.