Complete First problem in week 3 assignment
"In this problem you will convert an array of integers into a heap."

Since this was written in C#, no starter code was provided, so I was required to design an implement a rudimentary tree structure before beginning.

Please note that, due to the submission guidelines, my entire problem solution has to be contained in a single file.